% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_group_na}
\alias{count_group_na}
\title{Count Missing Values by group in a data frame}
\usage{
count_group_na(.data, .vars = dplyr::everything(), .group, .format = "longer")
}
\arguments{
\item{.data}{data frame}

\item{.vars}{variables to be counted.
If no variables were provided, the function will count Missing Values for all variables.}

\item{.group}{group variable}

\item{.format}{output data frame format ('longer' or 'wider')}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function Missing Values by group in a data frame
}
