% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_na}
\alias{count_na}
\title{Count Missing Values in a data frame}
\usage{
count_na(.data, .format = "longer")
}
\arguments{
\item{.data}{data frame}

\item{.format}{output data frame format ('longer' or 'wider')}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function counts Missing Values in a data frame
}
