% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna-package.R
\docType{package}
\name{dna-package}
\alias{dna}
\alias{dna-package}
\title{dna: Tools for handling Missing Data}
\description{
This package provides tools for handling Missing Data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arturhgq/dna}
  \item Report bugs at \url{https://github.com/arturhgq/dna/issues}
}

}
\author{
\strong{Maintainer}: Artur Quirino \email{contact.last@arturhgq.com} (\href{https://orcid.org/0000-0001-8071-6348}{ORCID})

}
\keyword{internal}
