% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{not_na}
\alias{not_na}
\title{In x there are no missing elements?}
\usage{
not_na(x)
}
\arguments{
\item{x}{an \R object to be tested}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function indicates which elements are not missing
}
