% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_na}
\alias{replace_na}
\title{Replace Missing Values by a value}
\usage{
replace_na(.data, ..., .value = 0)
}
\arguments{
\item{.data}{data frame}

\item{...}{variables. If no variables were provided, the function will fill
in missing values for all variables.}

\item{.value}{value for NA be replaced by}
}
\description{
Replace Missing Values by a value
}
