% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_na}
\alias{count_na}
\title{Count NAs in a data frame for each variable}
\usage{
count_na(data, format = "longer")
}
\arguments{
\item{data}{data frame}

\item{format}{output data frame format ('longer' or 'wider')}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function counts NAs in a data frame for each variable
}
\examples{
tibble::tibble(
var1 = c(1,2,3,4, NA),
var2 = c(NA, NA, NaN, 2, 3)
) -> data
count_na(data)
count_na(data, "wider")
}
