% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_na}
\alias{replace_na}
\title{Replace NAs by a value}
\usage{
replace_na(.data, ..., .value = 0)
}
\arguments{
\item{.data}{data frame}

\item{...}{variables. If no variables were provided, the function will fill
in Missing Values for all variables.}

\item{.value}{value for NAs be replaced by}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function replaces NAs by a value
}
\examples{
   tibble::tibble(
     var1 = c(1,2,3,4, NA),
     var2 = c(NA, NA, NaN, 2, 3)
   ) -> data
replace_na(data, .value = 0)
}
